tz-erl: Handle '/' in zone name even when there are no DST rules.
Previously the zonename function only split the Zone FORMAT field on
'/' if the current Rule specifies DST rules.  However it is possible
for the Zone entry to cover periods where there are Rule entries both
with and without DST rules.  Now zonename treats years without DST
rules as being on standard time for the treatment of the FORMAT field.